refactored in to separate helper
